Changed Subscriber to Google C++ Style coding
