+++ include/ROS2_Beginner_Tutorials/subscriber_func.hpp
@@ -1,0 +1,32 @@
+/**
+ * @file subscriber_func.hpp
+ * @author Vyshnav Achuthan (vyachu07@umd.edu)
+ * @brief 
+ * @version 0.1
+ * @date 2023-11-07
+ * 
+ * @copyright Copyright (c) 2023
+ * 
+ */
+#pragma once
+#include <memory>
+
+#include "rclcpp/rclcpp.hpp"
+#include "std_msgs/msg/string.hpp"
+using std::placeholders::_1;
+/**
+ * @brief Subscriber Class
+ * 
+ */
+class MinimalSubscriber : public rclcpp::Node
+{
+  public:
+    MinimalSubscriber()
+    : Node("minimal_subscriber")
+    {
+      subscription_ = this->create_subscription<std_msgs::msg::String>(
+      "topic", 10, std::bind(&MinimalSubscriber::topic_callback, this, _1));
+    }
+  private:
+      rclcpp::Subscription<std_msgs::msg::String>::SharedPtr subscription_;
+};
