+++ src/subscriber_func.cpp
@@ -8,38 +8,10 @@
  * @copyright Copyright (c) 2023
  * 
  */
+#include "../include/ROS2_Beginner_Tutorials/subscriber_func.hpp"
 
-#include <memory>
-
-#include "rclcpp/rclcpp.hpp"
-#include "std_msgs/msg/string.hpp"
-using std::placeholders::_1;
-/**
- * @brief Subscriber Class
- * 
- */
-class MinimalSubscriber : public rclcpp::Node
+void MinimalSubscriber::topic_callback(const std_msgs::msg::String & msg) const
 {
-  public:
-    MinimalSubscriber()
-    : Node("minimal_subscriber")
-    {
-      subscription_ = this->create_subscription<std_msgs::msg::String>(
-      "topic", 10, std::bind(&MinimalSubscriber::topic_callback, this, _1));
-    }
-
-  private:
-    void topic_callback(const std_msgs::msg::String & msg) const
-    {
       RCLCPP_INFO(this->get_logger(), "I heard: '%s'", msg.data.c_str());
-    }
-    rclcpp::Subscription<std_msgs::msg::String>::SharedPtr subscription_;
-};
-
-int main(int argc, char * argv[])
-{
-  rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<MinimalSubscriber>());
-  rclcpp::shutdown();
-  return 0;
 }
+
